Removed all the code to copy GitHubFlowVersion to the repository and refactored out the code to execute GitHubFlowVersion to a helper class Also added another Then clause to the empty repository acceptance test
